Another approach for zbuffering

+++ easy_image.h
@@ -230,7 +230,7 @@ namespace img
 			 * 	assert(y1 < getHeight())
 			 */
 			void draw_line(unsigned int x0, unsigned int y0, unsigned int x1, unsigned int y1, Color color);
-            void draw_zbuf_line(unsigned int x0, unsigned int y0, double z0, unsigned int x1, unsigned int y1, double z1, img::Color &color);
+            void draw_zbuf_line(double x0, double y0, double z0, double x1, double y1, double z1, img::Color &color);
             void draw_zbuf_triag(const Vector3D& A, const Vector3D& B, const Vector3D& C, double d, double dx, double dy, std::vector<double>& fullAmbientRef, std::vector<double>& diffuseRef, std::vector<double>& specularRef, double& refCoeff, const std::vector<Light*>& lights, const Vector3D& eyeCamera, const Matrix &eyeTransf, bool shadowOn, const std::vector<Texture*>& textures, const std::vector<Texture*>& fig_cubemap_textures, const std::vector<std::vector<double> >& uv, std::vector<Vector3D* > norm, const double& cubeSize);
             void draw_zbuf_triag_line(unsigned int x0, unsigned int y0, unsigned int x1, unsigned int y1, img::Color &color, double xg, double yg, double zg, double dzdx, double dzdy);
             void fillShadowBuffers(const std::vector<Figure>& figures, std::vector<Light*> &lights, double d, double dx, double dy) const;
